adds target patch system, that scans only matches in tha patch's time period

+++ db_models/registry/match_players.py
@@ -1,7 +1,7 @@
 from sqlalchemy import ForeignKeyConstraint, UniqueConstraint
 from sqlalchemy.orm import Mapped
 from sqlalchemy.orm import mapped_column
-from sqlalchemy.dialects.postgresql import BOOLEAN, CHAR, ENUM, INTEGER, SMALLINT
+from sqlalchemy.dialects.postgresql import BIGINT, BOOLEAN, CHAR, ENUM, SMALLINT
 
 from .base import Base
 from ..static import Regions, Roles
@@ -11,7 +11,7 @@ class MatchPlayers(Base):
     __tablename__ = "match_players"
 
     region: Mapped[Regions] = mapped_column(ENUM(Regions, create_type=False), primary_key=True)
-    match_id: Mapped[int] = mapped_column(INTEGER, primary_key=True)
+    match_id: Mapped[int] = mapped_column(BIGINT, primary_key=True)
     player_id: Mapped[str] = mapped_column(CHAR(78), primary_key=True)
     champion_id: Mapped[int] = mapped_column(SMALLINT, nullable=False)
     role: Mapped[Roles] = mapped_column(ENUM(Roles, create_type=False), nullable=False)
