+++ db_models/registry/players.py
@@ -1,7 +1,7 @@
 from sqlalchemy import UniqueConstraint
 from sqlalchemy.orm import Mapped
 from sqlalchemy.orm import mapped_column
-from sqlalchemy.dialects.postgresql import CHAR, SMALLINT, VARCHAR
+from sqlalchemy.dialects.postgresql import CHAR, INTEGER, VARCHAR
 
 from .base import Base
 
@@ -9,7 +9,7 @@ from .base import Base
 class Players(Base):
     __tablename__ = "players"
 
-    id: Mapped[int] = mapped_column(SMALLINT, primary_key=True)
+    id: Mapped[int] = mapped_column(INTEGER, primary_key=True)
     riot_id: Mapped[str] = mapped_column(CHAR(78), nullable=False)
     name: Mapped[str] = mapped_column(VARCHAR(16), nullable=False)
     tag: Mapped[str] = mapped_column(VARCHAR(5), nullable=False)
